Added byte, short, long and double options for numeric input

+++ include/libyam/expr.h
@@ -2,6 +2,7 @@
 #define EXPR_H_
 
 #include <stddef.h>
+#include <stdint.h>
 
 #define YAM_TOK_STD_TERM ':'
 
@@ -13,13 +14,14 @@ const char *yam_tok_next(const char *input, char term, size_t *parsed_len);
 const char *yam_tok_kv(const char *input, size_t len, const char *key,
                        size_t *parsed_len);
 
-const char *yam_tok_kv_adv(const char **input, size_t *len, const char *key, size_t *parsed_len);
+const char *yam_tok_kv_adv(const char **input, size_t *len, const char *key,
+                           size_t *parsed_len);
 
 // trims a token
 const char *yam_tok_trim(const char *tok, size_t *len);
 
 // converts a token to int
-int yam_tok_to_int(const char *tok, size_t len);
+int64_t yam_tok_to_int(const char *tok, size_t len);
 float yam_tok_to_float(const char *tok, size_t len);
 
 #endif
